Make particle viewer a switcher.
